Refatore projeto para C++ modular
