+++ .env.example
@@ -1,0 +1,8 @@
+# NET_IFACE=wlan0
+# RELAY_PIN=17
+# EMAIL_USER=seu_email@dominio
+# EMAIL_PASS=sua_senha
+# EMAIL_TO=destinatario@dominio
+# SMTP_SERVER=smtp.gmail.com
+# SMTP_PORT=587
+# EMAIL_USE_SSL=0
