+++ src/domain/MacProvider.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <optional>
+#include <string>
+
+struct MacProvider {
+    virtual ~MacProvider() = default;
+    virtual std::optional<std::string> next() = 0;
+};
