+++ src/domain/Relay.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+struct Relay {
+    virtual ~Relay() = default;
+    virtual bool pulse(double seconds, bool active_high = true) = 0;
+};
