+++ src/infra/JsSpeedTester.hpp
@@ -1,0 +1,4 @@
+#pragma once
+
+// Implementado em OoklaSpeedTester::runJs(). Este arquivo existe apenas para manter a
+// compatibilidade com a estrutura descrita no documento original.
