+++ src/util/Env.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <filesystem>
+#include <string>
+
+void loadDotenv(const std::filesystem::path& dotenvPath);
+std::string getenvOr(const std::string& key, const std::string& defVal);
