+++ src/util/Process.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <string>
+
+int runCmdCapture(const std::string& cmd, std::string& output);
+bool hasCmd(const std::string& name);
