+++ src/util/Time.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <string>
+
+std::string nowStr();
